backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
alembic==1.16.5
psycopg[binary]==3.2.10
python-jose[cryptography]==3.5.0
cryptography==46.0.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic-settings==2.10.1
python-multipart==0.0.20
celery==5.5.3
redis==6.4.0
openai==1.107.0
pypdf==6.0.0
python-docx==1.1.2
pandas==2.3.2
pgvector==0.4.1
httpx==0.28.1
email-validator==2.2.0
pytest==8.4.2
